build: Starry-Blog v3.0 正式发布

+++ blog-core/src/main/java/com/star/core/dto/CategoryDTO.java
@@ -24,6 +24,6 @@ public class CategoryDTO {
     /**
      * 分类下的文章数量
      */
-    private Long articleCount;
+    private Integer articleCount;
 
 }
